+++ Samples/iOS-Swift/iOS-Swift/ViewControllers/PerformanceViewController.swift
@@ -23,6 +23,11 @@ class PerformanceViewController: UIViewController {
         fatalError("init(coder:) has not been implemented")
     }
 
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+        SentrySDK.reportFullyDisplayed()
+    }
+
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
         startTest()